It's a Quiz about the Flags of different Countries.
In this App you have to first enter your name in the space given. The App will then ask you questions about various FLAGS and you will have to identify which Flag belongs to which country. Correct answer will be shown in GREEN and incorrect answer if will be shown in RED. Your total score will be displayed in the results screen after the QUIZ is finished.